refactor: extract FairBot business logic into SendChatCommand/Handler
Move all business logic from FairBotController (442 lines) into the
standard Controller → Command → Handler CQRS pattern:

- SendChatCommand: message, intentHint, contextPayload, requestId, user context
- SendChatValidator: FluentValidation for message length and required fields
- SendChatHandler: roster context enrichment, payload size limits,
  Agent Service proxy, error mapping (429/413/504/500)
- FairBotController: thin shell (~85 lines) with auth guard, legacy
  param resolution, request ID handling, mediator dispatch

Add Result<T> factory methods: Of413, Of429, Of504 for handler error mapping.
